application/pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Application for the Marketplace project"
dependencies = [
    "flask",
    "python-dotenv",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"